Address empty pcap raise errors
When an empty pcap is provided (one with valid headers and no packets)
the SortCap class will currently fail (python raise) when fd.close is
called in the close method. This happens typically when on writes or
reads are made since the fd is set lazily.

This new model isolates the write and read faciliies to separate classes
and ensures that we write a new pcap file even if the source is empty.